+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @aws-amplify/backend-function
 
+## 1.16.0
+
+### Minor Changes
+
+- 7d0ba5e: feat: add durable function configuration support
+- a34cd89: feat: add `description` prop to function schedule
+
+### Patch Changes
+
+- 7d0ba5e: chore: upgrade CDK dependencies
+- 4603f7a: bump aws-cdk-lib version to ^2.234.1 across all packages
+- f470ca7: Update lambda logging configuration to use non-deprecated AWS CDK properties
+- Updated dependencies [7d0ba5e]
+- Updated dependencies [299c804]
+- Updated dependencies [4603f7a]
+  - @aws-amplify/backend-output-storage@1.3.3
+  - @aws-amplify/plugin-types@1.11.2
+  - @aws-amplify/backend-output-schemas@1.8.0
+
 ## 1.15.2
 
 ### Patch Changes